% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidecar.R
\name{.sidecar_cache_dir}
\alias{.sidecar_cache_dir}
\title{Local cache directory for downloaded sidecars}
\usage{
.sidecar_cache_dir()
}
\description{
Local cache directory for downloaded sidecars
}
\keyword{internal}
