% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-lynker-spatial.R
\name{duckdb_connection}
\alias{duckdb_connection}
\title{Create a new DuckDB connection.}
\usage{
duckdb_connection(..., extensions = character(0))
}
\arguments{
\item{...}{Arguments passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{extensions}{Character vector of extensions to install and load on connect.}
}
\value{
A DBI connection to a DuckDB instance.
}
\description{
Create a new DuckDB connection.
}
\keyword{internal}
