% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store_filter_hl_reference}
\alias{store_filter_hl_reference}
\title{Filter network layer by hl_reference (supports pipe-delimited values)}
\usage{
store_filter_hl_reference(store, layer, hl_reference, ...)
}
\value{
A data.frame / tbl with matching rows
}
\description{
Filter network layer by hl_reference (supports pipe-delimited values)
}
\keyword{internal}
