% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store_get_layer}
\alias{store_get_layer}
\title{Get a store layer}
\usage{
store_get_layer(store, layer, ...)
}
\value{
Object inheriting from \code{tbl}
}
\description{
Get a store layer
}
\keyword{internal}
