% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store_has_layer}
\alias{store_has_layer}
\title{Check if a store has a layer}
\usage{
store_has_layer(store, layer, ...)
}
\value{
\code{logical(n)}
}
\description{
Check if a store has a layer
}
\keyword{internal}
