% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store_layer_cols}
\alias{store_layer_cols}
\title{Return column names for a store layer}
\usage{
store_layer_cols(store, layer, ...)
}
\value{
character vector of column names
}
\description{
Return column names for a store layer
}
\keyword{internal}
